Now works fairly well. Just demonstrates algorithm funcionality.
Robots will rotate until they find a transition from valid distance measurements (no blocks in the way) to invalid distance measurements (some blocks in the way), and wiggle around this transition. This should effectively shows where the algorithm can identify objects in the world other than the four walls. Obviously it doesn't work if the block that is in the way is too close to a wall, or if the robot itself is too close to a wall.

The actual point of this algorithm is to use this information to deduce the orientation of the robot, which is somethings the gps doesn't provide.