Now actually estimates its orientation correctly
The robots now use very simple integration of the wheel velocities in combination with the readings from the distance sensors to estimate their orientations. To illustrate this functionality, the robots are programmed to use this information to rotate themselves to point towards a point defined in the code called 'destination'. Work pretty well, although isn't super accurate: the integration of the wheel velocities obviously is quite crude and would alone lose calibration; the result of the algorithm applied to the distance sensor readings recalibrates the value whenever the distances sensors return valid distances (there aren't blocks/the other robot in the way), but the recalibration value itself isn't super accurate, especially if the robot is near a wall or is looking approximately orthogonally at the walls.